rpl_utility_server.cc(1359,46): warning C4267: 'argument': conversion from 'size_t' to 'unsigned int', possible loss of data
rpl_utility_server.cc(1359,46): error C2220: the following warning is treated as an error

+++ sql/log_event.h
@@ -4434,7 +4434,7 @@ public:
       @param[in] optional_metadata_len length of optional_metadata field.
      */
     Optional_metadata_fields(unsigned char* optional_metadata,
-                             unsigned int optional_metadata_len);
+                             size_t optional_metadata_len);
   };
 
   /**
